+++ src/classification/ast.parser.ts
@@ -1,12 +1,17 @@
-import { SourceFile, VariableDeclaration } from "ts-simple-ast";
+import { SourceFile, VariableDeclaration, Node, ObjectLiteralExpression, SyntaxKind, SyntaxList, Identifier } from "ts-simple-ast";
 import { Argumenter } from "@joejukan/argumenter";
+import { log } from "../function";
 
-export class ASTFinder {
+export class ASTParser {
     private source: SourceFile;
 
     public constructor(...args){
         let argue = new Argumenter(args);
         this.source = argue.instance(SourceFile);
+    }
+
+    public log(value: any){
+        log(`[ast-finder] ${value}`);
     }
 
     public findVariableDeclaration(type: {new ()}): VariableDeclaration;
@@ -41,5 +46,36 @@ export class ASTFinder {
             }
         }
         return undefined;
+    }
+    public exists(key: string, node: Node): boolean{
+        return typeof this.get(key, node) !== 'undefined';
+    }
+
+    public get(key: string, node: Node): ObjectLiteralExpression
+    public get(...args): ObjectLiteralExpression{
+        let argue = new Argumenter(args);
+        let key = argue.string;
+        let node:Node = <Node> argue.object;
+        let syntax:SyntaxList = undefined;
+
+        switch(node.getKind()){
+            case SyntaxKind.SyntaxList: 
+                syntax = <any> node;
+                break;
+            case SyntaxKind.ObjectLiteralExpression: 
+                syntax = node.getFirstChildByKind(SyntaxKind.SyntaxList)
+                break;
+        }
+
+        if(syntax){
+            let pairs = syntax.getChildren()
+            pairs.forEach(pair => {
+                let id = pair.getFirstChildByKind(SyntaxKind.Identifier);
+                if(key === id.getText() || `'${key}'` === id.getText() || `"${key}"` === id.getText())
+                    return pair.getFirstChildByKind(SyntaxKind.ObjectLiteralExpression);
+            });
+        }
+
+        return
     }
 }