+++ src/classification/index.ts
@@ -1,4 +1,4 @@
 export * from './ast.class';
-export * from './ast.finder';
+export * from './ast.parser';
 export * from './dependency.class';
 export * from './vue.di.plugin';