+++ src/function/index.ts
@@ -1,3 +1,4 @@
 export * from './function.string';
+export * from './log';
 export * from './pitch';
 export * from './pre.compile';