+++ src/globalization/variables.ts
@@ -1,2 +1,12 @@
 import { DependencyClass} from "../classification";
 export let dependencies: {[key: string]: DependencyClass} = {};
+export let configuration: { 
+    pitched?: boolean, 
+    verbose?: boolean,
+    deep?: boolean, 
+    entries?: string[]} = {
+    pitched: false,
+    verbose: false,
+    deep: true,
+    entries: new Array<string>()
+};